[RLlib] Remove TensorFlow-support + classes from new API stack. (#55042)
