+++ rllib/algorithms/dqn/default_dqn_rl_module.py
@@ -5,7 +5,6 @@ from ray.rllib.algorithms.sac.sac_learner import QF_PREDS
 from ray.rllib.core.columns import Columns
 from ray.rllib.core.learner.utils import make_target_network
 from ray.rllib.core.models.base import Encoder, Model
-from ray.rllib.core.models.specs.typing import SpecType
 from ray.rllib.core.rl_module.apis import QNetAPI, InferenceOnlyAPI, TargetNetworkAPI
 from ray.rllib.core.rl_module.rl_module import RLModule
 from ray.rllib.utils.annotations import (
@@ -140,7 +139,7 @@ class DefaultDQNRLModule(RLModule, InferenceOnlyAPI, TargetNetworkAPI, QNetAPI):
             return {}
 
     @override(RLModule)
-    def input_specs_train(self) -> SpecType:
+    def input_specs_train(self):
         return [
             Columns.OBS,
             Columns.ACTIONS,
@@ -148,15 +147,15 @@ class DefaultDQNRLModule(RLModule, InferenceOnlyAPI, TargetNetworkAPI, QNetAPI):
         ]
 
     @override(RLModule)
-    def output_specs_exploration(self) -> SpecType:
+    def output_specs_exploration(self):
         return [Columns.ACTIONS]
 
     @override(RLModule)
-    def output_specs_inference(self) -> SpecType:
+    def output_specs_inference(self):
         return [Columns.ACTIONS]
 
     @override(RLModule)
-    def output_specs_train(self) -> SpecType:
+    def output_specs_train(self):
         return [
             QF_PREDS,
             QF_TARGET_NEXT_PREDS,
