+++ rllib/algorithms/sac/default_sac_rl_module.py
@@ -8,7 +8,6 @@ from ray.rllib.algorithms.sac.sac_learner import (
 )
 from ray.rllib.core.learner.utils import make_target_network
 from ray.rllib.core.models.base import Encoder, Model
-from ray.rllib.core.models.specs.typing import SpecType
 from ray.rllib.core.rl_module.apis import InferenceOnlyAPI, QNetAPI, TargetNetworkAPI
 from ray.rllib.core.rl_module.rl_module import RLModule
 from ray.rllib.policy.sample_batch import SampleBatch
@@ -131,7 +130,7 @@ class DefaultSACRLModule(RLModule, InferenceOnlyAPI, TargetNetworkAPI, QNetAPI):
         return {}
 
     @override(RLModule)
-    def input_specs_train(self) -> SpecType:
+    def input_specs_train(self):
         return [
             SampleBatch.OBS,
             SampleBatch.ACTIONS,
@@ -139,7 +138,7 @@ class DefaultSACRLModule(RLModule, InferenceOnlyAPI, TargetNetworkAPI, QNetAPI):
         ]
 
     @override(RLModule)
-    def output_specs_train(self) -> SpecType:
+    def output_specs_train(self):
         return (
             [
                 QF_PREDS,
