+++ rllib/core/models/tests/test_catalog.py
@@ -31,24 +31,17 @@ from ray.rllib.core.models.torch.base import TorchModel
 from ray.rllib.core.rl_module.default_model_config import DefaultModelConfig
 from ray.rllib.core.rl_module.rl_module import RLModuleSpec
 from ray.rllib.models import MODEL_DEFAULTS
-from ray.rllib.models.tf.tf_distributions import (
-    TfCategorical,
-    TfDiagGaussian,
-    TfMultiCategorical,
-    TfMultiDistribution,
-)
 from ray.rllib.models.torch.torch_distributions import (
     TorchCategorical,
     TorchDiagGaussian,
     TorchMultiCategorical,
     TorchMultiDistribution,
 )
-from ray.rllib.utils.framework import try_import_tf, try_import_torch
+from ray.rllib.utils.framework import try_import_torch
 from ray.rllib.utils.numpy import convert_to_numpy
 from ray.rllib.utils.spaces.space_utils import get_dummy_batch_for_space
 from ray.rllib.utils.torch_utils import convert_to_torch_tensor
 
-_, tf, _ = try_import_tf()
 torch, _ = try_import_torch()
 
 
@@ -58,13 +51,11 @@ class TestCatalog(unittest.TestCase):
 
         Args:
             model: The model to check.
-            framework: The framework to use (tf|torch).
+            framework: The framework to use (torch).
             model_config_dict: The model config dict to use.
             input_space: The input space to use.
         """
-        convert_method = (
-            tf.convert_to_tensor if framework == "tf2" else convert_to_torch_tensor
-        )
+        convert_method = convert_to_torch_tensor
         expected_latent_dim = model_config_dict.get("latent_dim")
         if expected_latent_dim is None:
             # For CNNEncoders, `output_dims` are computed automatically.
@@ -132,7 +123,7 @@ class TestCatalog(unittest.TestCase):
             ),
         ]
 
-        frameworks = ["tf2", "torch"]
+        frameworks = ["torch"]
 
         # First check if encoders can be created for non-composite spaces
         print("Testing encoders for non-composite input spaces...")
@@ -182,10 +173,12 @@ class TestCatalog(unittest.TestCase):
             # Box
             TestConfig(
                 Box(-np.inf, np.inf, (7,), dtype=np.float32),
-                {"torch": TorchDiagGaussian, "tf2": TfDiagGaussian},
+                {
+                    "torch": TorchDiagGaussian,
+                },
             ),
             # Discrete
-            TestConfig(Discrete(5), {"torch": TorchCategorical, "tf2": TfCategorical}),
+            TestConfig(Discrete(5), {"torch": TorchCategorical}),
             # Nested Dict
             TestConfig(
                 Dict(
@@ -196,7 +189,6 @@ class TestCatalog(unittest.TestCase):
                 ),
                 {
                     "torch": TorchMultiDistribution,
-                    "tf2": TfMultiDistribution,
                 },
             ),
             # Nested Tuple
@@ -209,7 +201,6 @@ class TestCatalog(unittest.TestCase):
                 ),
                 {
                     "torch": TorchMultiDistribution,
-                    "tf2": TfMultiDistribution,
                 },
             ),
             # Tuple nested inside Dict
@@ -231,7 +222,6 @@ class TestCatalog(unittest.TestCase):
                 ),
                 {
                     "torch": TorchMultiDistribution,
-                    "tf2": TfMultiDistribution,
                 },
             ),
             # Dict nested inside Tuple
@@ -256,13 +246,12 @@ class TestCatalog(unittest.TestCase):
                 ),
                 {
                     "torch": TorchMultiDistribution,
-                    "tf2": TfMultiDistribution,
                 },
             ),
             # MultiDiscrete
             TestConfig(
                 MultiDiscrete([5, 5, 5]),
-                {"torch": TorchMultiCategorical, "tf2": TfMultiCategorical},
+                {"torch": TorchMultiCategorical},
             ),
         ]
 
@@ -284,10 +273,7 @@ class TestCatalog(unittest.TestCase):
 
             # Check if we can query the required input dimensions
             expected_cls = expected_cls_dict["torch"]
-            if (
-                expected_cls is TorchMultiDistribution
-                or expected_cls is TfMultiDistribution
-            ):
+            if expected_cls is TorchMultiDistribution:
                 # For these special cases, we need to create partials of the
                 # expected classes so that we can calculate the required inputs
                 expected_cls = _multi_action_dist_partial_helper(
@@ -295,10 +281,7 @@ class TestCatalog(unittest.TestCase):
                     action_space=action_space,
                     framework="torch",
                 )
-            elif (
-                expected_cls is TorchMultiCategorical
-                or expected_cls is TfMultiCategorical
-            ):
+            elif expected_cls is TorchMultiCategorical:
                 # For these special cases, we need to create partials of the
                 # expected classes so that we can calculate the required inputs
                 expected_cls = _multi_categorical_dist_partial_helper(
