+++ rllib/core/rl_module/rl_module.py
@@ -8,7 +8,6 @@ import gymnasium as gym
 
 from ray.rllib.core import DEFAULT_MODULE_ID
 from ray.rllib.core.columns import Columns
-from ray.rllib.core.models.specs.typing import SpecType
 from ray.rllib.core.rl_module.default_model_config import DefaultModelConfig
 from ray.rllib.models.distributions import Distribution
 from ray.rllib.utils.annotations import (
@@ -742,29 +741,29 @@ class RLModule(Checkpointable, abc.ABC):
         """
         return self
 
-    def output_specs_inference(self) -> SpecType:
+    def output_specs_inference(self):
         return [Columns.ACTION_DIST_INPUTS]
 
-    def output_specs_exploration(self) -> SpecType:
+    def output_specs_exploration(self):
         return [Columns.ACTION_DIST_INPUTS]
 
-    def output_specs_train(self) -> SpecType:
+    def output_specs_train(self):
         """Returns the output specs of the forward_train method."""
         return {}
 
-    def input_specs_inference(self) -> SpecType:
+    def input_specs_inference(self):
         """Returns the input specs of the forward_inference method."""
         return self._default_input_specs()
 
-    def input_specs_exploration(self) -> SpecType:
+    def input_specs_exploration(self):
         """Returns the input specs of the forward_exploration method."""
         return self._default_input_specs()
 
-    def input_specs_train(self) -> SpecType:
+    def input_specs_train(self):
         """Returns the input specs of the forward_train method."""
         return self._default_input_specs()
 
-    def _default_input_specs(self) -> SpecType:
+    def _default_input_specs(self):
         """Returns the default input specs."""
         return [Columns.OBS]
