+++ rllib/core/testing/bc_algorithm.py
@@ -6,11 +6,8 @@ to test the algorithm construction and customization.
 
 from ray.rllib.algorithms import Algorithm, AlgorithmConfig
 from ray.rllib.policy.torch_policy_v2 import TorchPolicyV2
-from ray.rllib.policy.eager_tf_policy_v2 import EagerTFPolicyV2
 from ray.rllib.core.testing.torch.bc_module import DiscreteBCTorchModule
 from ray.rllib.core.testing.torch.bc_learner import BCTorchLearner
-from ray.rllib.core.testing.tf.bc_module import DiscreteBCTFModule
-from ray.rllib.core.testing.tf.bc_learner import BCTfLearner
 from ray.rllib.core.rl_module.rl_module import RLModuleSpec
 from ray.rllib.utils.annotations import override
 from ray.rllib.utils.typing import ResultDict
@@ -23,14 +20,10 @@ class BCConfigTest(AlgorithmConfig):
     def get_default_rl_module_spec(self):
         if self.framework_str == "torch":
             return RLModuleSpec(module_class=DiscreteBCTorchModule)
-        elif self.framework_str == "tf2":
-            return RLModuleSpec(module_class=DiscreteBCTFModule)
 
     def get_default_learner_class(self):
         if self.framework_str == "torch":
             return BCTorchLearner
-        elif self.framework_str == "tf2":
-            return BCTfLearner
 
 
 class BCAlgorithmTest(Algorithm):
@@ -38,8 +31,6 @@ class BCAlgorithmTest(Algorithm):
     def get_default_policy_class(cls, config: AlgorithmConfig):
         if config.framework_str == "torch":
             return TorchPolicyV2
-        elif config.framework_str == "tf2":
-            return EagerTFPolicyV2
         else:
             raise ValueError("Unknown framework: {}".format(config.framework_str))
